requirements.txt
python-dotenv==1.0.1
selenium==4.26.1
webdriver_manager==4.0.2
